+++ src/cli/__init__.py
@@ -5,7 +5,6 @@ A CLI for controlling DC/OS clusters on Docker.
 import io
 import json
 import logging
-import shutil
 import subprocess
 import sys
 import tarfile
@@ -14,7 +13,7 @@ from ipaddress import IPv4Address
 from pathlib import Path
 from shutil import rmtree
 from subprocess import CalledProcessError
-from tempfile import gettempdir, gettempprefix
+from tempfile import gettempdir
 from typing import (  # noqa: F401
     Any,
     Callable,
@@ -49,6 +48,15 @@ from ._common import (
     VARIANT_LABEL_KEY,
     WORKSPACE_DIR_LABEL_KEY,
     existing_cluster_ids,
+)
+from ._doctor_checks import (
+    check_free_space,
+    check_memory,
+    check_mount_tmp,
+    check_networking,
+    check_ssh,
+    check_storage_driver,
+    link_to_troubleshooting,
 )
 from ._validators import (
     validate_cluster_exists,
@@ -135,6 +143,7 @@ def _is_enterprise(build_artifact: Path, workspace_dir: Path) -> bool:
     result = subprocess.check_output(
         args=get_version_args,
         cwd=str(workspace_dir),
+        stderr=subprocess.PIPE,
     )
 
     # In some cases, the name of the generated file is included in the output.
@@ -363,6 +372,7 @@ def create(
     base_workspace_dir = workspace_dir or Path(gettempdir())
     workspace_dir = base_workspace_dir / uuid.uuid4().hex
 
+    doctor_message = 'Try `dcos-docker doctor` for troubleshooting help.'
     ssh_keypair_dir = workspace_dir / 'ssh'
     ssh_keypair_dir.mkdir(parents=True)
     public_key_path = ssh_keypair_dir / 'id_rsa.pub'
@@ -380,6 +390,7 @@ def create(
         )
     except subprocess.CalledProcessError:
         rmtree(path=str(workspace_dir), ignore_errors=True)
+        click.echo(doctor_message)
         raise
 
     if enterprise:
@@ -422,13 +433,18 @@ def create(
         workspace_dir=workspace_dir,
     )
 
-    cluster = Cluster(
-        cluster_backend=cluster_backend,
-        masters=masters,
-        agents=agents,
-        public_agents=public_agents,
-        files_to_copy_to_installer=files_to_copy_to_installer,
-    )
+    try:
+        cluster = Cluster(
+            cluster_backend=cluster_backend,
+            masters=masters,
+            agents=agents,
+            public_agents=public_agents,
+            files_to_copy_to_installer=files_to_copy_to_installer,
+        )
+    except CalledProcessError as exc:
+        click.echo('Error creating cluster.', err=True)
+        click.echo(doctor_message)
+        sys.exit(exc.returncode)
 
     nodes = {
         *cluster.masters,
@@ -466,10 +482,10 @@ def create(
                 extra_config=extra_config,
             )
     except CalledProcessError as exc:
-        click.echo('Error creating cluster:', err=True)
-        click.echo(str(exc), err=True)
+        click.echo('Error installing DC/OS.', err=True)
+        click.echo(doctor_message)
         cluster.destroy()
-        return
+        sys.exit(exc.returncode)
 
     click.echo(cluster_id)
 
@@ -483,13 +499,17 @@ def list_clusters() -> None:
         click.echo(cluster_id)
 
 
-@dcos_docker.command('destroy')
+@dcos_docker.command('destroy-list')
 @click.argument(
     'cluster_ids',
     nargs=-1,
     type=str,
 )
-def destroy(cluster_ids: List[str]) -> None:
+@click.pass_context
+def destroy_list(
+    ctx: click.core.Context,
+    cluster_ids: List[str],
+) -> None:
     """
     Destroy clusters.
 
@@ -503,20 +523,32 @@ def destroy(cluster_ids: List[str]) -> None:
             click.echo(warning, err=True)
             continue
 
-        cluster_containers = _ClusterContainers(cluster_id=cluster_id)
-        containers = {
-            *cluster_containers.masters,
-            *cluster_containers.agents,
-            *cluster_containers.public_agents,
-        }
-        rmtree(path=str(cluster_containers.workspace_dir), ignore_errors=True)
-        for container in containers:
-            container.stop()
-            container.remove(v=True)
+        ctx.invoke(
+            destroy,
+            cluster_id=cluster_id,
+        )
 
-        client = docker.from_env(version='auto')
-        client.volumes.prune()
-        click.echo(cluster_id)
+
+@dcos_docker.command('destroy')
+@_existing_cluster_id_option
+def destroy(cluster_id: str) -> None:
+    """
+    Destroy a cluster.
+    """
+    cluster_containers = _ClusterContainers(cluster_id=cluster_id)
+    containers = {
+        *cluster_containers.masters,
+        *cluster_containers.agents,
+        *cluster_containers.public_agents,
+    }
+    rmtree(path=str(cluster_containers.workspace_dir), ignore_errors=True)
+    for container in containers:
+        container.stop()
+        container.remove(v=True)
+
+    client = docker.from_env(version='auto')
+    client.volumes.prune()
+    click.echo(cluster_id)
 
 
 class _ClusterContainers:
@@ -869,7 +901,7 @@ def sync_code(cluster_id: str, dcos_checkout_dir: str) -> None:
     """
     local_packages = Path(dcos_checkout_dir) / 'packages'
     local_test_dir = local_packages / 'dcos-integration-test' / 'extra'
-    if not local_test_dir.exists():
+    if not Path(local_test_dir).exists():
         message = (
             'DCOS_CHECKOUT_DIR must be set to the checkout of a DC/OS '
             'repository.\n'
@@ -923,164 +955,15 @@ def sync_code(cluster_id: str, dcos_checkout_dir: str) -> None:
         )
 
 
-def _info(message: str) -> None:
-    """
-    Show a warning message.
-    """
-    click.echo()
-    click.echo(click.style('Note: ', fg='blue'), nl=False)
-    click.echo(message)
-
-
-def _warn(message: str) -> None:
-    """
-    Show a warning message.
-    """
-    click.echo()
-    click.echo(click.style('Warning: ', fg='yellow'), nl=False)
-    click.echo(message)
-
-
-def _error(message: str) -> None:
-    """
-    Show an error message.
-    """
-    click.echo()
-    click.echo(click.style('Error: ', fg='red'), nl=False)
-    click.echo(message)
-
-
 @dcos_docker.command('doctor')
 def doctor() -> None:
     """
     Diagnose common issues which stop DC/OS E2E from working correctly.
     """
-    free_space = shutil.disk_usage(gettempdir()).free
-    free_space_gb = free_space / 1024 / 1024 / 1024
-
-    low_space_message = (
-        'The default temporary directory ("{tmp_prefix}") has '
-        '{free_space:.1f} GB of free space available. '
-        'Creating a cluster typically takes approximately 2 GB of '
-        'temporary storage. '
-        'If you encounter problems with disk space usage, set the '
-        '``TMPDIR`` environment variable to a suitable temporary '
-        'directory or use the ``--workspace-dir`` option on the '
-        '``dcos-docker create`` command.'
-    ).format(
-        tmp_prefix=Path('/') / gettempprefix(),
-        free_space=free_space_gb,
-    )
-
-    if free_space_gb < 5:
-        _warn(message=low_space_message)
-
-    client = docker.from_env(version='auto')
-    host_driver = client.info()['Driver']
-    docker_for_mac = bool(client.info()['OperatingSystem'] == 'Docker for Mac')
-    storage_driver_url = (
-        'https://docs.docker.com/storage/storagedriver/select-storage-driver/'
-    )
-    if host_driver not in DOCKER_STORAGE_DRIVERS:
-        message = (
-            "The host's Docker storage driver is \"{host_driver}\". "
-            'We recommend that you use one of: {supported_drivers}. '
-            'See {help_url}.'
-        ).format(
-            host_driver=host_driver,
-            supported_drivers=', '.join(sorted(DOCKER_STORAGE_DRIVERS.keys())),
-            help_url=storage_driver_url,
-        )
-        _warn(message)
-
-    if shutil.which('ssh') is None:
-        _error(message='`ssh` must be available on your path.')
-
-    # Image for a container which sleeps for a long time.
-    tiny_image = 'luca3m/sleep'
-
-    ping_container = client.containers.run(
-        image=tiny_image,
-        tty=True,
-        detach=True,
-    )
-
-    ping_container.reload()
-    ip_address = ping_container.attrs['NetworkSettings']['IPAddress']
-
-    try:
-        subprocess.check_call(
-            args=['ping', ip_address, '-c', '1', '-t', '1'],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-    except subprocess.CalledProcessError:
-        message = 'Cannot connect to a Docker container by its IP address.'
-        if docker_for_mac:
-            message += (
-                ' We recommend using '
-                'https://github.com/wojas/docker-mac-network. '
-            )
-        _error(message=message)
-
-    ping_container.stop()
-    ping_container.remove(v=True)
-
-    tmp_path = Path('/tmp').resolve()
-
-    try:
-        private_mount_container = client.containers.run(
-            image=tiny_image,
-            tty=True,
-            detach=True,
-            volumes={
-                str(tmp_path): {
-                    'bind': '/test',
-                },
-            },
-        )
-    except docker.errors.APIError as exc:
-        message = (
-            'There was an error mounting a the temporary directory path '
-            '"{tmp_path}" in container: \n\n'
-            '{exception_detail}'
-        ).format(
-            tmp_path=tmp_path,
-            exception_detail=exc.explanation.decode(
-                'ascii',
-                'backslashreplace',
-            ),
-        )
-        _error(message=message)
-    else:
-        private_mount_container.stop()
-        private_mount_container.remove(v=True)
-
-    docker_memory = client.info()['MemTotal']
-    message = (
-        'Docker has approximately {memory:.1f} GB of memory available. '
-        'The amount of memory required depends on the workload. '
-        'For example, creating large clusters or multiple clusters requires '
-        'a lot of memory.\n'
-        'A four node cluster seems to work well on a machine with 9 GB '
-        'of memory available to Docker.'
-    ).format(
-        memory=docker_memory / 1024 / 1024 / 1024,
-    )
-    mac_message = (
-        '\n'
-        'To dedicate more memory to Docker for Mac, go to '
-        'Docker > Preferences > Advanced.'
-    )
-    if docker_for_mac:
-        message += mac_message
-
-    _info(message=message)
-
-    message = (
-        'If you continue to experience problems, more information is '
-        'available at '
-        'http://dcos-e2e.readthedocs.io/en/latest/docker-backend.html#troubleshooting'  # noqa: E501
-    )
-
-    _info(message=message)
+    check_free_space()
+    check_storage_driver()
+    check_ssh()
+    check_networking()
+    check_mount_tmp()
+    check_memory()
+    link_to_troubleshooting()