+++ src/dcos_e2e/node.py
@@ -61,9 +61,10 @@ class Node:
         self,
         args: List[str],
         user: str,
-        env: Optional[Dict[str, Any]] = None,
-        shell: bool = False,
-        tty: bool = False,
+        env: Optional[Dict[str, Any]],
+        shell: bool,
+        tty: bool,
+        quiet: bool,
     ) -> List[str]:
         """
         Return a command to run `args` on this node over SSH.
@@ -74,14 +75,15 @@ class Node:
             env: Environment variables to be set on the node before running
                 the command. A mapping of environment variable names to
                 values.
-            shell: If False (the default), each argument is passed as a
-                literal value to the command.  If True, the command line is
-                interpreted as a shell command, with a special meaning applied
-                to some characters (e.g. $, &&, >). This means the caller must
-                quote arguments if they may contain these special characters,
-                including whitespace.
+            shell: If False, each argument is passed as a literal value to the
+                command. If True, the command line is interpreted as a shell
+                command, with a special meaning applied to some characters
+                (e.g. $, &&, >). This means the caller must quote arguments if
+                they may contain these special characters, including
+                whitespace.
             tty: If ``True``, allocate a pseudo-tty. This means that the users
                 terminal is attached to the streams of the process.
+            quiet: Whether the SSH command is to be run in quiet mode.
 
         Returns:
             The full SSH command to be run.
@@ -95,11 +97,13 @@ class Node:
         if tty:
             ssh_args.append('-t')
 
-        ssh_args += [
+        if quiet:
             # Suppress warnings.
-            # In particular, we don't care about remote host identification
-            # changes.
-            '-q',
+            # In particular, we often do not care about remote host
+            # identification changes.
+            ssh_args.append('-q')
+
+        ssh_args += [
             # This makes sure that only keys passed with the -i option are
             # used. Needed when there are already keys present in the SSH
             # key chain, which cause `Error: Too many Authentication
@@ -134,6 +138,7 @@ class Node:
         env: Optional[Dict[str, Any]] = None,
         shell: bool = False,
         tty: bool = False,
+        quiet: bool = True,
     ) -> subprocess.CompletedProcess:
         """
         Run a command on this node the given user.
@@ -157,6 +162,7 @@ class Node:
                 terminal is attached to the streams of the process.
                 This means that the values of stdout and stderr will not be in
                 the returned ``subprocess.CompletedProcess``.
+            quiet: If ``False``, show SSH warnings.
 
         Returns:
             The representation of the finished process.
@@ -174,6 +180,7 @@ class Node:
             env=env,
             shell=shell,
             tty=tty,
+            quiet=quiet,
         )
 
         return run_subprocess(
@@ -218,6 +225,7 @@ class Node:
             env=env,
             shell=shell,
             tty=False,
+            quiet=True,
         )
         return subprocess.Popen(
             args=ssh_args,